removed base coordinator class and removed support of <kafka-0.9
